1) Continue porting FT on GCM
2) Add a halfbodies node in each Runtime (singleton as the default node). This node is a standard one, but no active object can be created in it ; it only stores all Halfbodies created in this runtime. This allows to apply technical services also on Halfbodies.

Note that you can test if a Node or a nodeURL is the reserved Halfbodies node with NodeFactory.isHalfbodiesNode(...).

This selection is based on a particular node reserved name... Yes, I know, it seems ugly, but that's the only way. However, node names are now checked to not conflict with the halfbodies node reserved name (see NodeFactory). 

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/branches/SHORT_FTonGCM@8767 28e8926c-6b08-0410-baaa-805c5e19b8d6