Replaced almost all calls to RemoteObjectExposer(URI url) by RemoteObjectExposer(String name)
The goal is to only have to specify the name of the endpoint when binding an object. Each RemoteObjectFactory will decide the URL to use. It should be easier to implement the multi-protocol feature with the new method. Three calls have not been replaced since it will involve to modify the public API.



git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/branches/SHORT_Message_Routing@10988 28e8926c-6b08-0410-baaa-805c5e19b8d6