Solved the problem of object serialization and the use of it in message receivers. Active object registry is not need anymore (but still present in this commit).
Renamed some files

Resolved the problem of remote expositon.
For this purpose, added a service deployed at the server starting which
will deploy our active objects on the server.

Add commentaries.

Deleted PAAxisServlet class which is not needed anymore:
it was used to extend the AxisServlet class in order to be
able to get the axis configuration (which is a protected field).

Added the axis2-1.4 repository in which the ServiceDeployer service
is defined and that enables to create the axis2.war file to put into
the webapps directory of the http server.

If we want to use the local jetty server, we don't need to create this war
file since the Jetty server programmatically register an axis servlet.

Minnor changes

cleaned before committing

Now we can define the name we want for our service and
the methods argument is now taken into account.

Added component exposition.
Rename axis2-1.4 directory to axis2.
Change build.xml files in order to make it compilable typing "build proActiveWar" from the compile directory.

Forgotten to format

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/branches/SHORT_ffonteno_axis2_exportation@11586 28e8926c-6b08-0410-baaa-805c5e19b8d6