SpaceConfiguration: simplify getUrl()&getHostname(), del resolveHostname()
Clean up related code - move down url field to I/O configuration and
simplify computing logic.
We rather do not need explicit resolveHostname(). We probably even do not
need caching these results, so let's just compute it in place, as it is
usually used only one time.
The only drawback of such solution is misleading getHostname() and getUrl()
on non-target host, but who calls that. Even if somebody calls that, (s)he
has to handle ConfigurationException anyway,  so it is similar problem:
of awareness of ScratchSpaceConfiguration usage context.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/branches/ProActive_DataSpaces@12201 28e8926c-6b08-0410-baaa-805c5e19b8d6