explicit synchronization added
added in two places: after data spaces startup on each node - returned
ObjectsForSynchro futures are passed to waitForAll method of PAFutures
and after gathering URI's of scratches with partial results (which is
also passed to waitForAll).

Those URI's are collected into List (till now it was a Set), as waitForAll
requires List. For me it is quite strange but OK..

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/branches/ProActive_DataSpaces@12553 28e8926c-6b08-0410-baaa-805c5e19b8d6