Introduced new methods into Data Spaces API - resolve* with path parameter
Mostly one needs to open precise file in a data space and till now it had
to be done explicitly by resolving data space FileObject, creating
descendant FileObject for a path and calling create* method on it. All
FileObjects finally needed to be closed.

Since now this process is much easier and more comfortable.


git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/branches/ProActive_DataSpaces@12574 28e8926c-6b08-0410-baaa-805c5e19b8d6