HelloExample: revert to Utils.getHostname()
Last change to getLocalHost().getHostname() was looking hopefully, but it
made example stop working on my machine, as it involved using ssh and
required private key being not secured by password. SSH was being used,
because hostname in XML and hostname from Utils.getHostname() did *not*
match.

It appears that setting getLocalHost().getCanonicalHostname() would be
better if we want to achieve really suitable hostname for that. But
actually, if we want to force, that we set local hostname correctly, let's
use this Utils.getHostname() even if it has some side effects - correct
hostname is crucial to execute example in local environment. If
somebody want to use distributed environment it is still fine, he/she
just changes GCMD.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/branches/ProActive_DataSpaces@12883 28e8926c-6b08-0410-baaa-805c5e19b8d6