GCMApplicationImpl: move startDataSpaces() to constructor; issues remain
Moving it to constructor (as discussed on last meeting) allows us to use
DataSpacesTechnicalService on local bodies and half-bodies nodes, and
set up or access some spaces before starting deployment.

However, still we require correct application id on this nodes to be set
up. In current implementation, app id is deployment id, which is not set
up on local nodes.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/branches/ProActive_DataSpaces@13009 28e8926c-6b08-0410-baaa-805c5e19b8d6