Extract NodeScratchSpace interface + improve docs of *ScratchSpace
NodeScratchSpace was the last class in .core package that was directly
dependent on VFS library. Now, class from .core rely only on
implementations (of classes, interfaces) from our .vfs package.

As a side effect, providing initialization arguments was moved from
constructor to init() method, which seems to be ok.

Documentation of NodeScratchSpace and ApplicationScratchSpace was improved.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/branches/ProActive_DataSpaces@13141 28e8926c-6b08-0410-baaa-805c5e19b8d6