PADataSpaces: change getCapabilitiesForSpaceType() to constants
They are changed to constants for 2 reasons:
1) SpaceType is an element of .core package, and was the only element
from .core that user had to see. This was also the only occurrence of
SpaceType in .api.
2) When user refers to this static method and provides constant (enum) as
an argument, he/she is actually referring to constant.
If we want to have it dynamic, we should provide
getCapabilitiesForFileObjectSpaceType() or getCapabilitiesForFileObject()
(not technically possible).

DataSpacesImpl.checkCapabiltiesOrWound() required changes.

BTW, constants are now wrapped with unmodifiable set.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/branches/ProActive_DataSpaces@13145 28e8926c-6b08-0410-baaa-805c5e19b8d6