FileSystemServerDeployer: move to parent package, add there VFS URL support
As we move FileSystemServerDeployer to .vfsprovider package, adding
dependencies on .client, .server, .protocol packages is not a problem.
So, ProActiveFileName.getServerVFSRootURL() method call can be there now,
making deployer easier/less strange to use.

Deployer seems to be also the "main" class of .vfsprovider, possibly used
by user, so root package looks ok for it.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/branches/ProActive_DataSpaces@13511 28e8926c-6b08-0410-baaa-805c5e19b8d6