PROACTIVE-709: configure DataSpaces for local nodes (bodies', half-bodies')
Now, local nodes are configured in GCMA, so DataSpaces can be used from
AOs on default node or even from non-AOs (half-bodies node).

This feature brings some limitation: only one GCM application with
DataSpaces can be configured on one runtime. Subsequent DataSpaces
configurations are ignored.
It is like that because otherwise it would cause an issue with identifying
proper application.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/branches/ProActive_DataSpaces@13519 28e8926c-6b08-0410-baaa-805c5e19b8d6