Solved the problem of namespace in cxf webservices: Forced elementFormDefault to be unqualified for services as well as for clients.
Added interceptors which are useless and will be removed when refactoring code

Changed configuration to set java.rmi.server.RMIClassLoaderSpi to org.objectweb.proactive.core.classloading.protocols.ProActiveRMIClassLoader.
Without this configuration, we cannot use webservices with deployment (the ServiceDeployer cannot unmarshalled the object to expose as a WS because
of an empty codebase).

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/branches/SHORT_Axis2_CXF@14217 28e8926c-6b08-0410-baaa-805c5e19b8d6