Web Services: Changed lib location and added some loggers
Web services libs (whether it be axis2 or cxf one) are located into the lib directory, in two distinct directories.
For common libs, another directory as been created. This enables in particular to remove double libraries.

When building proctive, all these libraries are copied to ProActive/dist/lib at the same level and inserted as indexes into
ProActive.jar. This enables to remove the addition of cxf-manifest.jar to the CLASSPATH in launching scripts.

Compilation of webservices is now divided by two: one for each framework.
Thus, we know which libraries are needed for each framework.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/branches/SHORT_Axis2_CXF@14224 28e8926c-6b08-0410-baaa-805c5e19b8d6